refactor: calculateRate to use strategy pattern
